Safari-609.1.11.1/LayoutTests/accessibility/misspelling-range-expected.txt

wrods is misspelled aab lotsi nowadays. euep.
This tests that misspelling ranges are properly retrieved in the fashion that a spell checker would.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


textMarkerRange start: 0
textMarkerRange end: 45
startRange start: 0
startRange end: 0
misspelling start: 0
misspelling end: 5
PASS text.stringForTextMarkerRange(misspellingRange) is 'wrods'
misspelling start: 20
misspelling end: 23
PASS text.stringForTextMarkerRange(misspellingRange) is 'aab'
misspelling start: 24
misspelling end: 29
PASS text.stringForTextMarkerRange(misspellingRange) is 'lotsi'
misspelling start: 40
misspelling end: 44
PASS text.stringForTextMarkerRange(misspellingRange) is 'euep'
misspelling start: 24
misspelling end: 29
PASS text.stringForTextMarkerRange(misspellingRange) is 'lotsi'
PASS successfullyParsed is true

TEST COMPLETE
